+++ src/routes/owner/admins/$id/index.ts
@@ -1,0 +1,40 @@
+import { Hono } from 'hono'
+
+import type { AppContext } from '@/context'
+
+import { requestValidator } from '@/middleware'
+
+import {
+  cancelAdminInviteHandler,
+  getAdminHandler,
+  resendAdminInviteHandler,
+  updateAdminHandler,
+} from './handler'
+import { adminIdParamsSchema, updateAdminBodySchema } from './schema'
+
+export const ownerAdminByIdRoute = new Hono<AppContext>()
+
+ownerAdminByIdRoute.get(
+  '/',
+  requestValidator('param', adminIdParamsSchema),
+  getAdminHandler,
+)
+
+ownerAdminByIdRoute.patch(
+  '/',
+  requestValidator('param', adminIdParamsSchema),
+  requestValidator('json', updateAdminBodySchema),
+  updateAdminHandler,
+)
+
+ownerAdminByIdRoute.post(
+  '/resend-invite',
+  requestValidator('param', adminIdParamsSchema),
+  resendAdminInviteHandler,
+)
+
+ownerAdminByIdRoute.post(
+  '/cancel-invite',
+  requestValidator('param', adminIdParamsSchema),
+  cancelAdminInviteHandler,
+)
